requirements.txt
aiogram==3.3.0
environs==10.3.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
flake8==7.0.0
flake8-annotations-complexity==0.0.8
flake8_bugbear==24.2.6
flake8_builtins==2.3.0
flake8_commas==2.1.0
flake8_comprehensions==3.14.0
flake8_docstrings==1.7.0
flake8_eradicate==1.5.0
flake8_import_order==0.18.2
flake8_pep3101==2.1.0
flake8_print==5.0.0
flake8_rst_docstrings==0.3.0
flake8_string_format==0.3.0
flake8_variables_names==0.0.6